fix: 回收站内文件夹概率出现无法清空的情况.
遇到权限的问题时，回收站文件就会出现删除失败的情况。采用gio接口去删除回收站文件，并对回收站文件操作进行了限制处理，只有回收站根目录下的文件才能执行删除以及剪切操作

Log: 修复回收站文件无法删除的问题
Bug: https://pms.uniontech.com/zentao/bug-view-61677.html

Former-commit-id: 62388f2e347f60977f3813222ae9caad2caebd46

+++ src/dde-file-manager-lib/io/dlocalfilehandler.cpp
@@ -21,11 +21,18 @@
 #include "dlocalfilehandler.h"
 #include "private/dfilehandler_p.h"
 #include "../controllers/vaultcontroller.h"
+#include "dfmstandardpaths.h"
 
 #include <QObject>
 #include <QDir>
 #include <QDateTime>
 #include <QTemporaryFile>
+
+#undef signals
+extern "C" {
+#include <gio/gio.h>
+}
+#define signals public
 
 #include <unistd.h>
 #include <utime.h>
@@ -116,6 +123,9 @@ bool DLocalFileHandler::remove(const DUrl &url)
 {
     Q_ASSERT(url.isLocalFile());
 
+    if (url.isTrashFile() || url.path().startsWith(DFMStandardPaths::location(DFMStandardPaths::TrashFilesPath)))
+        return removeTrashFile(url);
+
     if (::remove(url.toLocalFile().toLocal8Bit()) == 0)
         return true;
 
@@ -129,6 +139,9 @@ bool DLocalFileHandler::remove(const DUrl &url)
 bool DLocalFileHandler::rmdir(const DUrl &url)
 {
     Q_ASSERT(url.isLocalFile());
+
+    if (url.isTrashFile() || url.path().startsWith(DFMStandardPaths::location(DFMStandardPaths::TrashFilesPath)))
+        return removeTrashFile(url);
 
     if (::rmdir(url.toLocalFile().toLocal8Bit()) == 0)
         return true;
@@ -191,6 +204,17 @@ bool DLocalFileHandler::setFileTime(const DUrl &url, const QDateTime &accessDate
     d->setErrorString(QString::fromLocal8Bit(strerror(errno)));
 
     return false;
+}
+
+bool DLocalFileHandler::removeTrashFile(const DUrl &url)
+{
+    const QString &filePath = url.toLocalFile();
+    const QString &trashFile = filePath.right(filePath.length() - DFMStandardPaths::location(DFMStandardPaths::TrashFilesPath).length()).prepend("trash://");
+
+    GFile *file = g_file_new_for_uri(trashFile.toStdString().c_str());
+    g_file_delete(file, nullptr, nullptr);
+    g_object_unref(file);
+    return true;
 }
 
 DFM_END_NAMESPACE
